+++ R/segment_decimation.R
@@ -11,8 +11,9 @@
 barycentric_predecimation = function(las, params = default_arbor_parameters)
 {
   res <- params$decimation$barycentric_predecimation_resolution
-  keep <- C_voxel_barycenter_decimate(las@data$X, las@data$Y, las@data$Z, res)
+  keep <- C_homogeneization(las@data, res, hybrid = FALSE)
   las@data$decimated <- keep
+  cat("Final retention ", sum(keep), " points (", round(sum(keep)/length(keep)*100, 1), "%)\n", sep = "")
   free(keep)
   return(las)
 }
@@ -20,13 +21,12 @@ barycentric_predecimation = function(las, params = default_arbor_parameters)
 barycentric_decimation = function(las, res)
 {
   decimated <- TRUE
-  keep <- C_voxel_barycenter_decimate(las@data$X, las@data$Y, las@data$Z, res)
-  las@data$decimated <- keep
-  las <- lidR::filter_poi(las, decimated == TRUE)
-  las@data$decimated = NULL
+  keep <- C_homogeneization(las@data, res, hybrid = FALSE)
+  las <- las[keep]
   free(keep)
   return(las)
 }
+
 #' Homogenization of the Point Cloud
 #'
 #' Homogenization of the point cloud using a hybrid approach that includes
@@ -38,16 +38,12 @@ barycentric_decimation = function(las, res)
 hybrid_homogeneization = function(las, res = 0.02)
 {
   decimated <- TRUE
-  keep <- C_voxel_barycenter_decimate(las@data$X, las@data$Y, las@data$Z, res)
-  n = sum(keep)
-  rm = which(!keep)
-  i = sample(rm, min((n*0.1), length(rm)))
-  keep[i] = TRUE
+  keep <- C_homogeneization(las@data, res)
   las = las[keep]
+  cat("Final retention ", sum(keep), " points (", round(sum(keep)/length(keep)*100, 1), "%)\n", sep = "")
   free(keep)
   return(las)
 }
-
 
 get_barycentric_predecimation <- function(las, params = default_arbor_parameters)
 {